Keep the reference to the message that delivered image list to test.
The message id of the message that was used by AppDB to send the image
list is kept for further reference and reported back to AppDB as part of
the outcome (final report or error message).

Message id is stored in the OpenNebula tempate that is created for the
image list so it's uniquely bound. The message id is added to a
generated template that is used by the cloudkeeper to create new
template records. In order to avoid race conditions (when the wront
message id would be used to register a template), Secant is advised to
use its own instance of Cloudkeeper, where it controls the process of
template generation. A locking mechanism was introduced to prevent
multiple instances of the Secant argo consumer from running.